ingestion/requirements_venv.txt
python-dotenv==1.0.0
pymilvus
sentence-transformers==2.2.2
numpy==1.24.2
pandas==1.5.3
huggingface_hub==0.23.2
